Split out settings plugin
